requirements.txt
setuptools
pyvisa
numpy
future
paramiko
pyserial
tk